[database] change accounting consumption field type
Set the "consumption" field of the table "accounting" to 'bigint'.

Fixes #41

+++ sources/core/database/pg_structure.sql
@@ -13,7 +13,7 @@ CREATE TABLE accounting (
   accounting_project varchar(255) NOT NULL default '',
   queue_name varchar(100) NOT NULL default '',
   consumption_type varchar(5) check (consumption_type in ('ASKED','USED')) NOT NULL default 'ASKED',
-  consumption integer NOT NULL default '0',
+  consumption bigint NOT NULL default '0',
   PRIMARY KEY  (window_start,window_stop,accounting_user,accounting_project,queue_name,consumption_type)
 );
 CREATE INDEX accounting_user ON accounting (accounting_user);
